# [PERF] hr_expense: add missing index `account.move.line.expense_id`