Deleting an expense is slow when the `account_move_line` table is large, as the `Many2one`/`fkey` `expense_id` needs to be set to `NULL` where necessary. If there is no index on the `fkey`, it's `Seq.Scan` on the table `account_move_line`.

| Before | After |
|--------|-------|
| 10s    | 58ms  |

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr